feat(example): move handshake code into mod
